The CustomRoom plaque now shows the room gametype.
Also moved the extension to get enum name into the enum script itself.
Also took possible language change into account.
